refact: generalize List checkbox component to become apart of the design system
- Generalizing the List checkbox component
- Adapting services list field in user dialog form to the new component system for List checkbox
