+++ app/dashboard/src/components/UserDialog/ServicesField.tsx
@@ -1,8 +1,14 @@
-import { FormControl, FormLabel, FormErrorMessage, Center, Spinner } from '@chakra-ui/react';
+import { FormControl, FormLabel, FormErrorMessage, Center, Spinner, Text, Spacer, Flex, chakra } from '@chakra-ui/react';
+import { UsersIcon, ServerStackIcon } from '@heroicons/react/24/outline';
+import { IconBaseStyle } from 'constants/IconBaseStyle';
 import { TFunction } from 'i18next';
 import { Controller } from 'react-hook-form';
 import { Service } from 'types/Service';
+import { ServicePropertyStatus } from './ServicePropertyStatus';
 import { ServicesSelectList } from './ServicesSelectList';
+
+const UsersServiceIcon = chakra(UsersIcon, IconBaseStyle);
+const InboundsServiceIcon = chakra(ServerStackIcon, IconBaseStyle);
 
 interface ServicesFieldProps {
   form: any;
@@ -27,7 +33,19 @@ export const ServicesField = ({ form, t, services }: ServicesFieldProps) => {
           return (
             <ServicesSelectList
               list={services}
-              {...field}
+              renderCheckboxContent={(service) => (
+                <>
+                  <Text fontSize="sm" color="gray.700" fontWeight="medium">
+                    {service.name}
+                  </Text>
+                  <Spacer />
+                  <Flex flexDirection="row" alignItems="center">
+                    <ServicePropertyStatus value={service.users.length} StatusIcon={UsersServiceIcon} />
+                    <Text color="primary.600" mx="8px">/</Text>
+                    <ServicePropertyStatus value={service.inbounds.length} StatusIcon={InboundsServiceIcon} />
+                  </Flex>
+                </>
+              )}
             />
           );
         }}
